requirements.txt
arxiv==1.4.5
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0
pandas>=2.0.0,<2.3.0
tabulate==0.9.0
google-genai>=1.28.0
numpy==2.2.4
python-dotenv==1.1.0
Requests==2.32.3
tenacity==8.3.0
